requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
certifi==2024.8.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via wandb
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
docker-pycreds==0.4.0
    # via wandb
filelock==3.16.1
    # via
    #   torch
    #   triton
fonttools==4.54.1
    # via matplotlib
fsspec==2024.10.0
    # via torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
kiwisolver==1.4.7
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.2
    # via
    #   seaborn
    #   ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.1.3
    # via
    #   zindilacunamalaria (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pandas
    #   scipy
    #   seaborn
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
opencv-python==4.10.0.84
    # via
    #   zindilacunamalaria (pyproject.toml)
    #   ultralytics
packaging==24.2
    # via matplotlib
pandas==2.2.3
    # via
    #   zindilacunamalaria (pyproject.toml)
    #   seaborn
    #   ultralytics
pillow==11.0.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
platformdirs==4.3.6
    # via wandb
protobuf==5.28.3
    # via wandb
psutil==6.1.0
    # via
    #   ultralytics
    #   wandb
py-cpuinfo==9.0.0
    # via ultralytics
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   zindilacunamalaria (pyproject.toml)
    #   ultralytics
    #   wandb
requests==2.32.3
    # via
    #   ultralytics
    #   wandb
scipy==1.14.1
    # via
    #   zindilacunamalaria (pyproject.toml)
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
sentry-sdk==2.18.0
    # via wandb
setproctitle==1.3.3
    # via wandb
setuptools==75.3.0
    # via wandb
six==1.16.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sympy==1.13.1
    # via torch
torch==2.5.1
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.20.1
    # via ultralytics
tqdm==4.67.0
    # via ultralytics
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   torch
    #   wandb
tzdata==2024.2
    # via pandas
ultralytics==8.3.28
    # via zindilacunamalaria (pyproject.toml)
ultralytics-thop==2.0.11
    # via ultralytics
urllib3==2.2.3
    # via
    #   requests
    #   sentry-sdk
wandb==0.18.6
    # via zindilacunamalaria (pyproject.toml)
